+++ src/cli/pile/blob.rs
@@ -101,7 +101,7 @@ pub fn run(cmd: Command) -> Result<()> {
             let res = (|| -> Result<(), anyhow::Error> {
                 let file_handle = File::open(&file)?;
                 let bytes = unsafe { Bytes::map_file(&file_handle)? };
-                let handle = pile.put::<UnknownBlob, _>(bytes)?;
+                let handle = pile.put(bytes)?;
                 let hash: triblespace_core::value::Value<Hash<Blake3>> = Handle::to_hash(handle);
                 let string: String = hash.from_value();
                 println!("{string}");