+++ src/cli/store/blob.rs
@@ -95,7 +95,7 @@ pub fn run(cmd: Command) -> Result<()> {
             let mut remote: ObjectStoreRemote<Blake3> = ObjectStoreRemote::with_url(&url)?;
             let file_handle = File::open(&file)?;
             let bytes = unsafe { Bytes::map_file(&file_handle)? };
-            let handle = remote.put::<UnknownBlob, _>(bytes)?;
+            let handle = remote.put(bytes)?;
             let hash: triblespace_core::value::Value<Hash<Blake3>> = Handle::to_hash(handle);
             let string: String = hash.from_value();
             println!("{string}");